Fix for title option
A variable is considered empty if it does not exist or if its value equals FALSE. Therefore, setting the title option to false was not having the desired effect of removing the title from the append/prepend components. Changed function to isset that determines if a variable is set and is not NULL. To summarize: empty(false) return true but !isset(false) returns false.